requirements.txt
flask
pyTelegramBotAPI
pytesseract
Pillow
gunicorn
